feat(bots): build factory-template service bots directly via openapi create
## Problem

openapi create-as-service (bot_type=service + engine_properties) 被 #1403 一期的组合
闸门限制为 personal-only,于是 #1911 用"翻译成 personal 创建 + 同秒自动
upgrade_to_service"绕行。绕行的第二步 BaasService.upgrade_bot 在 BaaS 开户(记录
PENDING)后的秒级窗口内调用 /api/v1/bots/{uuid}/update,稳定 404,且失败在
fire-and-forget daemon 线程里被吞——2xx 响应让机器编排永远等不到 online binding,
只能整链重建(2026-09-10 prod:同一 owner 26 分钟建 7 个孤儿 bot;prepub 另一接入方
同踩;人肉两步式因天然退避从不受影响)。

## Solution

让第二次 BaaS 调用消失: factory snapshot 模板(personalCoding/coship_new 等,带
template_key+template_uid 完整身份)的 coding create 直接以 bot_type=service 落到
create——一次 BaaS 调用(service 形态容器、BCN provider 注册、草稿发布单,create 链
:1410/:1732/:1873 均已支持),退役 service_intake 翻译层。

边界: 手写 applicationCoding(无工厂身份)保持 personal-only 409——其
workspace-hosting(分配+软删回滚)只打通过 personal 形态(#1403 一期),不半支持。

- strategy.py: 闸门 bot_type in {personal, service};legacy 手写分支新增 service 409
- combo_policy.py: 镜像同步(注释说明 factory/legacy 细化在 strategy 层)
- create_flow/create_context: service_intake/seam 通道退役(as_payload 兼容旧 key)
- openapi_v1 router: _FacadeServiceIntakeSeam/facade 注入删除,create 与 auth-status
  docstring 改直建语义;errors_bot_create 502 映射移除
- gateway bots.openapi.json: create description 手工单行同步

## Validation

- 新语义矩阵: test_create_flow_coding_service_create 9 用例(直建/personal 不变/
  legacy 409/pending 不创建;ISSUED 完成直建)
- 反转旧语义守卫: application_coding gates(factory+service 现在通过)、combo_policy、
  openapi endpoints(is_direct/无 convert;502 与 replay 用例随语义退役)
- 受影响面: openapi_v1 1645 passed / bot_management 1098 / bot_inventory 86 /
  service_bot 1056 / adapters bot_management 4;ruff clean
- prepub 验收(后续): 用复现样本 create-as-service+personalCoding 重放,期望一次
  baas_device_create(bot_type=service)、无 update 调用、无 Failed to restart bot
