+++ src/gateway/configs/schemas/bots.openapi.json
@@ -24133,7 +24133,7 @@
         ]
       },
       "post": {
-        "description": "Create a bot (201), or 202 with authorization URLs when consent is needed.\n\nOn a 202, have the user complete authorization at one of the returned\nURLs, then poll the auth-status endpoint — the bot is only created there,\non ISSUED. Engine-specific creation properties belong under\n\"engine_properties\" and are interpreted by the selected engine;\nengine configuration is managed through the engine-config endpoints after\ncreation.\n\nA coding template with bot_type=service (create-as-service) is fulfilled\norchestratively: the bot is created through the personal-only template path\nand immediately upgraded to a service bot, which is what the 201 reports. If\nthe upgrade fails the response is a 502 naming the created-as-personal bot —\nretry the lifecycle upgrade instead of re-creating.",
+        "description": "Create a bot (201), or 202 with authorization URLs when consent is needed.\n\nOn a 202, have the user complete authorization at one of the returned\nURLs, then poll the auth-status endpoint — the bot is only created there,\non ISSUED. Engine-specific creation properties belong under\n\"engine_properties\" and are interpreted by the selected engine;\nengine configuration is managed through the engine-config endpoints after\ncreation.\n\nA coding template with bot_type=service (create-as-service) is fulfilled\ndirectly: the strategy gate admits service for factory-snapshot template\ncreates, so the bot is provisioned in its service shape (one BaaS call —\nservice-shaped container, BCN registration, draft publish record) with no\nfollow-up upgrade. Hand-written application-coding configs stay\npersonal-only and answer 409 for bot_type=service.",
         "operationId": "create_bot_openapi_v1_bots_post",
         "parameters": [
           {
